Use API and Implementation file structure
