+++ GoogleSignIn/Sources/GIDAppCheck/API/GIDAppCheckProvider.h
@@ -14,34 +14,25 @@
  * limitations under the License.
  */
 
-#import <TargetConditionals.h>
 #import <Foundation/Foundation.h>
-#import "GoogleSignIn/Sources/Public/GoogleSignIn/GIDSignIn.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
 #if TARGET_OS_IOS && !TARGET_OS_MACCATALYST
-
+@protocol GIDAppCheckTokenFetcher;
 @class FIRAppCheckToken;
 
-/// A list of potential error codes returned from the Google Sign-In SDK during App Check.
-typedef NS_ERROR_ENUM(kGIDSignInErrorDomain, GIDAppCheckErrorCode) {
-  /// An unexpected error was encountered.
-  kGIDAppCheckUnexpectedError = 1,
-  /// `GIDAppCheck` has already performed the key generation and attestation steps.
-  kGIDAppCheckAlreadyPrepared = 2,
-};
-
-NS_CLASS_AVAILABLE_IOS(14)
-@interface GIDAppCheck : NSObject
+/// Interface providing the API for both pre-warming `GIDSignIn` to use Firebase App Check and
+/// fetching the App Check token.
+NS_AVAILABLE_IOS(14)
+@protocol GIDAppCheckProvider <NSObject>
 
 /// Creates the instance of this App Check wrapper class.
 ///
 /// If `provider` is nil, then we default to `FIRAppCheck`.
 ///
-/// @param provider The instance performing the Firebase App Check requeests.
-- (instancetype)initWithAppCheckProvider:(nullable id<GIDAppCheckProvider>)provider
-    NS_DESIGNATED_INITIALIZER;
+/// @param tokenFetcher The instance performing the Firebase App Check token requests.
+- (instancetype)initWithAppCheckTokenFetcher:(nullable id<GIDAppCheckTokenFetcher>)tokenFetcher;
 
 /// Prewarms the library for App Check by asking Firebase App Check to generate the App Attest key
 /// id and perform the initial attestation process (if needed).
@@ -61,9 +52,9 @@ NS_CLASS_AVAILABLE_IOS(14)
 /// Whether or not the App Attest key ID created and the attestation object has been fetched.
 @property(atomic, readonly, getter=isPrepared) BOOL prepared;
 
+
 @end
 
 #endif // TARGET_OS_IOS && !TARGET_OS_MACCATALYST
 
 NS_ASSUME_NONNULL_END
-
