+++ GoogleSignIn/Sources/GIDAppCheck/UI/GIDActivityIndicatorViewController.h
