+++ GoogleSignIn/Sources/GIDAppCheck/UI/GIDActivityIndicatorViewController.m
