+++ GoogleSignIn/Sources/GIDSignIn_Private.h
@@ -46,9 +46,6 @@ typedef void (^GIDDisconnectCompletion)(NSError *_Nullable error);
 /// Redeclare |currentUser| as readwrite for internal use.
 @property(nonatomic, readwrite, nullable) GIDGoogleUser *currentUser;
 
-/// Expose `GIDAppCheck` for tests.
-@property(nonatomic, readonly, nullable) GIDAppCheck *appCheck;
-
 /// Whether or not the sign in or add scopes flows should use the Firebase App Check token
 @property(nonatomic, readonly) BOOL useAppCheckToken;
